// web/src/app/api/model/reflection-response.ts
/**
 * Simplicity training API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ReflectionResponse { 
    body?: string;
    createdAt?: string;
    id?: string;
    title?: string;
    updatedAt?: string;
}